requirements.txt
Django>=4.2.1
djangorestframework>=3.13.1
djongo>=1.3.5